cleanup docker and add better flags for bash scripts

+++ script/docker.sh
@@ -22,33 +22,57 @@ make_container() {
 	docker exec -it "$CONTAINER_NAME" /bin/bash
 }
 
-case "$1" in
-	""|-d|--dev)
-		echo "Starting docker in profile dev"
-		make_container "dev"
-		;;
+if [[ $# -eq 0 ]]; then
+    print_help
+    exit 1
+fi
 
-	-p|--prod)
-		echo "Starting docker in profile prod"
-		make_container "prod"
-		;;
+while test $# -gt 0; do
+	case "$1" in
+		-d|--dev)
+			shift
 
-	-c|--clean)
-		echo "Cleaning container and image"
-		docker stop $(docker ps -aq)
-		docker rm $(docker ps -aq)
-		docker rmi -f $(docker images -q)
-		;;
+			if test $# -gt 0; then
+				echo "--dev cannot have more than one flag after it"
+				exit 1
+      fi
 
-	-h|--help)
-		print_help
-		;;
+			echo "Starting docker in profile dev"
+			make_container "dev"
+			;;
 
-	*)
-		echo "Unknown option: $1" >&2
-		echo "Use --help for usage."
-		exit 1
-		;;
-esac
+		-p|--prod)
+			shift
 
-exit 0
+			if test $# -gt 0; then
+				echo "--prod cannot have more than one flag after it"
+				exit 1
+      fi
+
+			echo "Starting docker in profile prod"
+			make_container "prod"
+			;;
+
+		-c|--clean)
+			shift
+
+			echo "Cleaning container and image"
+
+			docker stop $(docker ps -aq)
+			docker rm $(docker ps -aq)
+			docker rmi -f $(docker images -q)
+			;;
+
+		-h|--help)
+			shift
+			print_help
+			exit 1
+			;;
+
+		*)
+			echo "Unknown option: $1" >&2
+			echo "Use --help for usage."
+			exit 1
+			;;
+	esac
+done
